// frontend/src/app/api/models/event-type.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export type EventType = 'Vorgemerkt' | 'Beworben' | 'ScreeningGespraech' | 'ChallengeErhalten' | 'ChallengeAbgegeben' | 'Interview' | 'Kennenlerntag' | 'AngebotErhalten' | 'AngebotAngenommen' | 'AngebotAbgelehnt' | 'Absage' | 'Zurueckgezogen' | 'KeineRueckmeldung';